refactor(components): moves animation logic to taost | hj | #2159
