+++ packages/components/src/components/toast/toast/toast-manager.ts
@@ -1,6 +1,6 @@
 import { forceUpdate } from '@stencil/core';
 import type { ToastState } from './toast-utils';
-import { ANIMATION_DURATION } from '../../banner/banner-styles-shared';
+import { ANIMATION_DURATION } from './toast-styles';
 
 const TOAST_DEFAULT_TIMEOUT = 6000;
