+++ packages/components/src/components/toast/toast/toast-styles.spec.ts
@@ -1,4 +1,34 @@
-import { getComponentCss } from './toast-styles';
+import {
+  getAnimationIn,
+  getAnimationOut,
+  getComponentCss,
+  getKeyframesMobile,
+  KeyframesDirection,
+} from './toast-styles';
+
+describe('getAnimationIn()', () => {
+  it.each<string[]>([
+    ['mobileIn', '--p-duration'],
+    ['mobileIn', null],
+  ])('should return correct JssStyle for keyframesName: %s and durationVar: %s', (keyframesName, durationVar) => {
+    expect(getAnimationIn(keyframesName, durationVar)).toMatchSnapshot();
+  });
+});
+
+describe('getAnimationOut()', () => {
+  it('should return correct JssStyle', () => {
+    expect(getAnimationOut('mobileOut')).toMatchSnapshot();
+  });
+});
+
+describe('getKeyframesMobile()', () => {
+  it.each<[KeyframesDirection, string]>([
+    ['in', '--p-bottom-var'],
+    ['out', '--p-bottom-var'],
+  ])('should return correct JssStyle for direction: %s and bottomVar: %s', (direction, bottomVar) => {
+    expect(getKeyframesMobile(direction, bottomVar)).toMatchSnapshot();
+  });
+});
 
 describe('getComponentCss()', () => {
   it('should return correct css', () => {
